pyproject.toml
[project]
name = "diaz-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "apache-airflow>=3.1.0",
    "asyncpg>=0.30.0",
    "bs4>=0.0.2",
    "celery[redis]>=5.5.3",
    "chromadb>=1.3.0",
    "ddgs>=9.6.1",
    "docling>=2.60.0",
    "duckdb>=1.4.1",
    "elasticsearch>=9.2.0",
    "faiss-cpu>=1.12.0",
    "fastapi[standard]>=0.120.2",
    "gunicorn>=23.0.0",
    "ipywidgets>=8.1.8",
    "jmespath>=1.0.1",
    "jq>=1.10.0",
    "langchain>=1.0.2",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.1",
    "langchain-openai>=1.0.1",
    "langchain-qdrant>=1.1.0",
    "langgraph>=1.0.2",
    "langgraph-checkpoint-mongodb>=0.2.1",
    "langgraph-checkpoint-postgres>=3.0.0",
    "langgraph-cli[inmem]>=0.4.5",
    "more-itertools>=10.8.0",
    "pandas>=2.3.3",
    "psycopg[binary,pool]>=3.2.12",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "pymongo>=4.15.3",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.15.1",
    "sqlalchemy>=2.0.45",
]
